# _examples/booktest/api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "booktest",
    "description": "Boilerplate code generated by **sqlc-grpc**. Modify _proto/*.proto_ files then run `buf generate` to change the services interface.",
    "version": "1.0",
    "contact": {
      "name": "sqlc-grpc",
      "url": "https://github.com/walterwanderley/sqlc-grpc"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/authors": {
      "post": {
        "operationId": "BooksService_CreateAuthor",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Author"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAuthorRequest"
            }
          }
        ],
        "tags": [
          "BooksService"
        ]
      }
    },
    "/authors/{authorId}": {
      "get": {
        "operationId": "BooksService_GetAuthor",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Author"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BooksService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Author": {
      "type": "object",
      "properties": {
        "authorId": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "v1CreateAuthorRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
  }
}